Rename file to keep it consistent
