+++ src/queries/specialized-queries/index.js
@@ -3,4 +3,4 @@
 exports.MoreLikeThisQuery = require('./more-like-this-query');
 exports.ScriptQuery = require('./script-query');
 exports.PercolateQuery = require('./percolate-query');
-exports.DistanceFeatureQuery = require('./distance-feature');
+exports.DistanceFeatureQuery = require('./distance-feature-query');
